move Mailboxes and Scheduled in the Adapter namespace
