+++ src/Adapter/Scheduled.php
@@ -1,11 +1,9 @@
 <?php
 declare(strict_types = 1);
 
-namespace Innmind\Witness;
+namespace Innmind\Witness\Adapter;
 
-use Innmind\Witness\{
-    Actor\Address\Name,
-};
+use Innmind\Witness\Actor\Address\Name;
 use Innmind\OperatingSystem\OperatingSystem;
 use Innmind\Immutable\Maybe;
